修复: log.js 模板字符串语法错误导致页面崩溃

+++ src/main/java/com/xxx/scm/service/impl/SysRoleServiceImpl.java
@@ -3,6 +3,7 @@ package com.xxx.scm.service.impl;
 import com.xxx.scm.entity.SysRole;
 import com.xxx.scm.exception.BusinessException;
 import com.xxx.scm.mapper.SysRoleMapper;
+import com.xxx.scm.mapper.SysUserMapper;
 import com.xxx.scm.service.SysRoleService;
 import com.xxx.scm.util.PageResult;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +16,9 @@ public class SysRoleServiceImpl implements SysRoleService {
 
     @Autowired
     private SysRoleMapper roleMapper;
+
+    @Autowired
+    private SysUserMapper userMapper;
 
     @Override
     public List<SysRole> findAll() {
@@ -51,10 +55,14 @@ public class SysRoleServiceImpl implements SysRoleService {
 
     @Override
     public void delete(Integer roleId) {
-        // 检查是否有用户关联了这个角色，防止删了正在用的角色
-        // 这里简单处理，经理角色(roleId=1)不能删
+        // 经理角色(roleId=1)不能删除
         if (roleId == 1) {
             throw new BusinessException("经理角色不能删除");
+        }
+        // 检查是否有用户关联了这个角色
+        int userCount = userMapper.countByRoleId(roleId);
+        if (userCount > 0) {
+            throw new BusinessException("该角色下还有 " + userCount + " 个用户，请先移除用户后再删除角色");
         }
         roleMapper.deleteById(roleId);
     }
